dcrnn att

+++ Models/dcrnn_att/dcrnn_cell_att.py
@@ -110,7 +110,7 @@ class DCGRUCell(RNNCell):
             self._supports.append(self._build_sparse_matrix(support))
 
         self._bias_mt = tf.convert_to_tensor(utils.adj_to_bias(np.expand_dims(adj_mx, axis=0),
-                                                               [self._num_nodes], nhood=1))
+                                                               [self._num_nodes], nhood=1), dtype=tf.float32)
         _adj_mx = tf.convert_to_tensor(adj_mx)
         self._adj_mx_repeat = tf.tile(tf.expand_dims(_adj_mx, axis=0), [batch_size, 1, 1])
         for support in self._supports:
